fix dvbapi stops working after reader restart
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@2285 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ oscam.c
@@ -2458,15 +2458,14 @@ int process_input(uchar *buf, int l, int timeout)
 
 static void restart_clients()
 {
-	  int i;
-	  cs_log("restarting clients");
-	  for (i=0; i<CS_MAXPID; i++) {
-	    if (client[i].typ=='c')
-	    {
-	      kill(client[i].pid, SIGKILL);
-	      cs_log("killing client c%02d pid %d", i, client[i].pid);
-	    }
-	  }
+	int i;
+	cs_log("restarting clients");
+	for (i=0; i<CS_MAXPID; i++) {
+		if (client[i].pid && client[i].typ=='c' && ph[client[i].ctyp].type & MOD_CONN_NET) {
+			kill(client[i].pid, SIGKILL);
+			cs_log("killing client c%02d pid %d", i, client[i].pid);
+		}
+	}
 }
